+++ src/lintrans/gui/dialogs/settings.py
@@ -100,6 +100,13 @@ class DisplaySettingsDialog(SettingsDialog):
         )
         self.dict_checkboxes['r'] = self.checkbox_draw_transformed_grid
 
+        self.checkbox_draw_basis_vectors = QCheckBox(self)
+        self.checkbox_draw_basis_vectors.setText('Draw basis &vectors')
+        self.checkbox_draw_basis_vectors.setToolTip(
+            'Draw the transformed basis vectors'
+        )
+        self.dict_checkboxes['v'] = self.checkbox_draw_basis_vectors
+
         # Animations
 
         self.checkbox_smoothen_determinant = QCheckBox(self)
@@ -161,6 +168,7 @@ class DisplaySettingsDialog(SettingsDialog):
         self.vlay_groupbox_basic_stuff.setSpacing(20)
         self.vlay_groupbox_basic_stuff.addWidget(self.checkbox_draw_background_grid)
         self.vlay_groupbox_basic_stuff.addWidget(self.checkbox_draw_transformed_grid)
+        self.vlay_groupbox_basic_stuff.addWidget(self.checkbox_draw_basis_vectors)
 
         self.groupbox_basic_stuff = QGroupBox('Basic stuff', self)
         self.groupbox_basic_stuff.setLayout(self.vlay_groupbox_basic_stuff)
@@ -206,6 +214,7 @@ class DisplaySettingsDialog(SettingsDialog):
         # Basic stuff
         self.checkbox_draw_background_grid.setChecked(self.display_settings.draw_background_grid)
         self.checkbox_draw_transformed_grid.setChecked(self.display_settings.draw_transformed_grid)
+        self.checkbox_draw_basis_vectors.setChecked(self.display_settings.draw_basis_vectors)
 
         # Animations
         self.checkbox_smoothen_determinant.setChecked(self.display_settings.smoothen_determinant)
@@ -223,6 +232,7 @@ class DisplaySettingsDialog(SettingsDialog):
         # Basic stuff
         self.display_settings.draw_background_grid = self.checkbox_draw_background_grid.isChecked()
         self.display_settings.draw_transformed_grid = self.checkbox_draw_transformed_grid.isChecked()
+        self.display_settings.draw_basis_vectors = self.checkbox_draw_basis_vectors.isChecked()
 
         # Animations
         self.display_settings.smoothen_determinant = self.checkbox_smoothen_determinant.isChecked()