+++ src/lintrans/gui/plots/widgets.py
@@ -62,7 +62,8 @@ class VisualizeTransformationWidget(VectorGridPlot):
         if self.display_settings.draw_transformed_grid:
             self.draw_transformed_grid(painter)
 
-        self.draw_basis_vectors(painter)
+        if self.display_settings.draw_basis_vectors:
+            self.draw_basis_vectors(painter)
 
         if self.display_settings.draw_eigenlines:
             self.draw_eigenlines(painter)